refactor: DateCell 위치 component 하위로 이동
